Changed to getError Added new ociGetError Moved helpers to cHelpers.go
